Move theming to separate module
