+++ auto_neutron/hub.py
@@ -5,13 +5,12 @@ from __future__ import annotations
 
 import atexit
 import csv
-import functools
 import logging
 import typing as t
 from functools import partial
 
 import babel
-from PySide6 import QtCore, QtGui, QtWidgets
+from PySide6 import QtCore, QtWidgets
 
 import auto_neutron.locale
 
@@ -19,6 +18,7 @@ import auto_neutron.locale
 from __feature__ import snake_case, true_property  # noqa: F401
 from auto_neutron import Theme, settings
 from auto_neutron.constants import JOURNAL_PATH, ROUTE_FILE_NAME, get_config_dir
+from auto_neutron.dark_theme import set_theme
 from auto_neutron.fuel_warn import FuelWarn
 from auto_neutron.game_state import PlotterState
 from auto_neutron.route_plots import AhkPlotter, CopyPlotter, NeutronPlotRow
@@ -259,69 +259,3 @@ class Hub(QtCore.QObject):
                 writer.writerows(row.to_csv() for row in self.plotter_state.route)
 
             settings.General.last_route_index = self.plotter_state.route_index
-
-
-class _ThemeSelector:
-    """Allow changing the app's theme through the `set_theme` method."""
-
-    # The palettes are cached because of a bug where reapplying the same dark palette but as a new QPalette instance
-    # caused the app's style to revert back to the default instead of being unchanged.
-
-    def set_theme(self, is_dark: bool) -> None:
-        """Set the app's theme depending on `is_dark`."""
-        if is_dark:
-            self._app.set_palette(self._dark_palette)
-        else:
-            self._app.set_palette(self._light_palette)
-
-    @functools.cached_property
-    def _dark_palette(self) -> QtGui.QPalette:
-        """Create a dark themed palette."""
-        p = QtGui.QPalette()
-        p.set_color(QtGui.QPalette.Window, QtGui.QColor(35, 35, 35))
-        p.set_color(QtGui.QPalette.WindowText, QtGui.QColor(247, 247, 247))
-        p.set_color(QtGui.QPalette.Base, QtGui.QColor(25, 25, 25))
-        p.set_color(QtGui.QPalette.Text, QtGui.QColor(247, 247, 247))
-        p.set_color(QtGui.QPalette.Button, QtGui.QColor(60, 60, 60))
-        p.set_color(QtGui.QPalette.AlternateBase, QtGui.QColor(45, 45, 45))
-        p.set_color(QtGui.QPalette.ToolTipText, QtCore.Qt.white)
-        p.set_color(QtGui.QPalette.ButtonText, QtCore.Qt.white)
-        p.set_color(QtGui.QPalette.PlaceholderText, QtGui.QColor(110, 110, 100))
-        p.set_color(QtGui.QPalette.Link, QtGui.QColor(0, 123, 255))
-        p.set_color(
-            QtGui.QPalette.Disabled,
-            QtGui.QPalette.Light,
-            QtGui.QColor(0, 0, 0),
-        )
-        p.set_color(
-            QtGui.QPalette.Disabled,
-            QtGui.QPalette.Text,
-            QtGui.QColor(110, 110, 100),
-        )
-        p.set_color(
-            QtGui.QPalette.Disabled,
-            QtGui.QPalette.ButtonText,
-            QtGui.QColor(110, 110, 100),
-        )
-        p.set_color(
-            QtGui.QPalette.Disabled,
-            QtGui.QPalette.Button,
-            QtGui.QColor(50, 50, 50),
-        )
-
-        return p
-
-    @functools.cached_property
-    def _light_palette(self) -> QtGui.QPalette:
-        """Create a light themed palette."""
-        p = self._app.style().standard_palette()
-        p.set_color(QtGui.QPalette.Link, QtGui.QColor(0, 123, 255))
-        return p
-
-    @functools.cached_property
-    def _app(self) -> QtWidgets.QApplication:
-        """Get the application's instance."""
-        return QtWidgets.QApplication.instance()
-
-
-set_theme = _ThemeSelector().set_theme
